requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pydantic[email]>=2.10.4
pydantic-settings>=2.7.1
supabase>=2.11.0
python-jose[cryptography]==3.3.0
cloudinary>=1.41.0
httpx>=0.28.1
python-dotenv>=1.0.1
setuptools>=75.1.0
passlib==1.7.4
bcrypt>=4.0.1
